refactor(cli): remove json format output
